Make a FASTA using UniProt accessions

Given a vector of UniProt accessions, this function will:

Download the sequences

Print the current UniProt release (you should put this in the FASTA file name)

(Optional) add cRAP numbers to the FASTA headers for each sequence

Save the sequences in a FASTA file

This function takes a Qfeatures object with the evidence.txt output from MaxQuant and filters it to retain peptides with a given PTM.

Usage

filter_maxquant_ptm(obj, i, ptm_prob_col = "Phospho..STY..Probabilities")

filter_maxquant_ptm(obj, i, ptm_prob_col = "Phospho..STY..Probabilities")

Arguments

obj

QFeatures. Proteomics dataset

i

string. Index for the SummarizedExperiment with data without imputation

ptm_prob_col

character column name for PTMs.

accessions

character vector, the UniProt accessions to use

file

character, filepath to save the fasta to e.g. "crap.fasta"

is_crap

logical, Is the output going to be a cRAP database? If TRUE cRAP001, cRAP002, etc. is appended to the sequence headers in the FASTA file. Default is FALSE

overwrite

logical, if the FASTA file already exists should it be overwritten? Default is FALSE

verbose

logical, should the function send messages to the console? Default is TRUE

Value

Returns a FASTA file saved to disk at the specified file path.

Returns a SummarizedExperiment where the value for the PTM probability column is not empty