Filter Proteome Discoverer DDA output

This function filters the output .txt files (peptide groups or PSMs) from Proteome Discoverer for DDA, based on various criteria:

Remove features without a master protein

(Optional) Remove features without a unique master protein (i.e. Number.of.Protein.Groups == 1)

(Optional) Remove features matching a cRAP protein

(Optional) Remove features matching any protein associated with a cRAP protein (see below)

Remove features without quantification values (only if TMT or SILAC are TRUE and level = "peptide".)

Usage

filter_features_pd_dda(
  obj,
  master_protein_col = "Master.Protein.Accessions",
  protein_col = "Protein.Accessions",
  unique_master = TRUE,
  silac = FALSE,
  TMT = FALSE,
  level = "peptide",
  filter_crap = TRUE,
  crap_proteins = NULL,
  filter_associated_crap = TRUE,
  remove_no_quant = TRUE
)

Arguments

obj

SummarisedExperiment containing output from Proteome Discoverer. Use readQFeatures to read in .txt file

master_protein_col

string. Name of column containing master proteins.

protein_col

string. Name of column containing all protein matches.

unique_master

logical. Filter out features without a unique master protein.

silac

logical. Is the experiment a SILAC experiment?

TMT

logical. Is the experiment a TMT experiment?

level

string. Type of input file, must be one of either "peptide" or "PSM".

filter_crap

logical. Filter out features which match a cRAP protein.

crap_proteins

character vector. The protein IDs form the cRAP proteins,

filter_associated_crap

logical. Filter out features which match a cRAP associated protein.

remove_no_quant

logical. Remove features with no quantification

Value

Returns a SummarisedExperiment with the filtered Proteome Discoverer output.

Details

Associated cRAP proteins are proteins which have at least one feature shared with a cRAP protein. It has been observed that the cRAP fasta files often do not contain all possible cRAP proteins e.g. some features can be assigned to a keratin which is not in the provided cRAP database.

In the example below, using filter_associated_crap = TRUE will filter out f2 and f3 in addition to f1, regardless of the value in the Master.Protein.Accession column.

feature  Protein.Accessions         Master.Protein.Accessions
f1       protein1, protein2, cRAP,  protein1,
f2       protein1, protein3         protein3,
f3       protein2                   protein2

Examples

if (FALSE) { # \dontrun{

#### PSMs.txt example ####
# load PD PSMs.txt output
tmt_qf <- readQFeatures(assayData = psm_tmt_total,
 quantCols = 36:45,
 name = "psms_raw")

# extract the UniProt accessions from the cRAP FASTA headers
crap_accessions <- get_crap_fasta_accessions(crap_fasta_inf)

# filter thg peptides
psm2 <- filter_features_pd_dda(
  obj = tmt_qf[['psms_raw']],
  master_protein_col = "Master.Protein.Accessions",
  protein_col = "Protein.Accessions",
  unique_master = TRUE,
  TMT = TRUE,
  level = "PSM",
  filter_crap = TRUE,
  crap_proteins = crap_accessions,
  filter_associated_crap = TRUE
)

} # }